pyproject.toml
[project]
name = "uchroma"
version = "1.1.0"
description = "Userspace RGB control for Razer Chroma peripherals"
readme = "README.md"
license = "LGPL-3.0-only"
requires-python = ">=3.10"
authors = [{ name = "Stefanie Kondik", email = "stef@hyperbliss.tech" }]
keywords = ["razer", "chroma", "rgb", "keyboard", "mouse", "linux"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
dependencies = [
    "argcomplete",
    "coloraide>=4.0",
    "colorlog",
    "dbus-fast>=2.0",
    "evdev",
    "frozendict",
    "hid",
    "hsluv",
    "numpy",
    "pyudev",
    "ruamel.yaml",
    "scikit-image",
    "traitlets",
    "wrapt",
]

[project.optional-dependencies]
gtk = [
    "pygobject>=3.50",
]

[project.scripts]
uchroma = "uchroma.client.client:run_client"
uchromad = "uchroma.server.server:run_server"
uchroma-gtk = "uchroma.gtk:main"

[project.entry-points."uchroma.plugins"]
renderers = "uchroma.fxlib"

[build-system]
requires = ["setuptools>=61", "cython>=3.0", "numpy"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "uchroma",
    "uchroma.fxlib",
    "uchroma.client",
    "uchroma.server",
    "uchroma.gtk",
    "uchroma.gtk.models",
    "uchroma.gtk.widgets",
    "uchroma.gtk.panels",
    "uchroma.gtk.services",
    "uchroma.gtk.resources",
]
include-package-data = true

[tool.setuptools.package-data]
"uchroma.server" = ["data/*.yaml"]
"uchroma.gtk.resources" = ["*.css"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
    "ty>=0.0.1a7",
    "cython>=3.0",
    "pygobject-stubs>=2.16.0",
]

# ─────────────────────────────────────────────────────────────────────────────
# Ruff - Linting & Formatting
# ─────────────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
    "C4",     # flake8-comprehensions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "PLC",    # pylint conventions
    "PLE",    # pylint errors
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E741",   # ambiguous variable name
    "UP007",  # use X | Y for union types (not always desired)
    "SIM108", # use ternary (not always clearer)
]

[tool.ruff.lint.isort]
known-first-party = ["uchroma"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PT"]  # pytest style not enforced in tests
# D-Bus files use D-Bus type signatures ('s', 'u', 'as', 'a{sv}') as annotations - not Python types
"**/dbus.py" = ["F821", "F722"]
"**/dbus_*.py" = ["F821", "F722"]

# ─────────────────────────────────────────────────────────────────────────────
# ty - Type Checking
# ─────────────────────────────────────────────────────────────────────────────
[tool.ty.environment]
python-version = "3.10"

[tool.ty.rules]
# Ignore missing stubs for external packages (hid, pyudev, evdev, coloraide, dbus-fast)
unresolved-import = "ignore"
# Traitlets uses dynamic attributes extensively
unresolved-attribute = "ignore"
possibly-missing-attribute = "ignore"
# Gradual adoption - warn on these
possibly-unresolved-reference = "warn"
invalid-type-form = "warn"
# Errors we want to fix
invalid-assignment = "error"
invalid-argument-type = "warn"
invalid-return-type = "warn"

[tool.ty.src]
include = ["uchroma/**/*.py"]
exclude = ["**/build/**", "**/.venv/**"]

# Relax rules for D-Bus files that use D-Bus type signatures as annotations
[[tool.ty.overrides]]
include = ["**/dbus.py", "**/dbus_*.py"]
[tool.ty.overrides.rules]
unresolved-reference = "ignore"
invalid-return-type = "ignore"
invalid-syntax-in-forward-annotation = "ignore"
